pyproject.toml
[build-system]
requires = ["scikit-build-core", "numpy", "scipy"]  # Include dependencies for building
build-backend = "scikit_build_core.build"

[tool.scikit-build]
cmake.source-dir = 'spades/radial'
cmake.build-type = "Release"
cmake.minimum-version = "3.15"
build.verbose = true
logging.level = "INFO"
sdist.cmake = true
build-dir = "build"

[project]
name = "spades"
version = "0.1.0"
description = "sajkdgh"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "numpy",
    "scipy",
    "pyyaml",
    "hepunits",
    "periodictable",
    "tqdm",
    "numba",
    "mpmath",
]

[project.optional-dependencies]
docs = ["sphinx>=7.0"]

[project.scripts]
spades = "spades.bin.compute_spectra_psfs:main"

[project.urls]
"Homepage" = "https://github.com/yourusername/spades"
"Documentation" = "https://yourdocs.example.com"
"Source" = "https://github.com/yourusername/spades"
